+++ acmed/src/acme_proto/http.rs
@@ -1,7 +1,8 @@
 use crate::acme_proto::structs::{AcmeError, ApiError, Directory, HttpApiError};
 use crate::certificate::Certificate;
+use crate::rate_limits;
 use acme_common::error::Error;
-use http_req::request::{Method, Request};
+use http_req::request::{self, Method};
 use http_req::response::Response;
 use http_req::uri::Uri;
 use std::path::Path;
@@ -10,6 +11,12 @@ use std::{thread, time};
 
 const CONTENT_TYPE_JOSE: &str = "application/jose+json";
 const CONTENT_TYPE_JSON: &str = "application/json";
+
+struct Request<'a> {
+    r: request::Request<'a>,
+    uri: &'a Uri,
+    method: Method,
+}
 
 struct DummyString {
     pub content: String,
@@ -25,13 +32,7 @@ impl FromStr for DummyString {
     }
 }
 
-fn new_request<'a>(
-    cert: &Certificate,
-    root_certs: &'a [String],
-    uri: &'a Uri,
-    method: Method,
-) -> Request<'a> {
-    cert.debug(&format!("{}: {}", method, uri));
+fn new_request<'a>(root_certs: &'a [String], uri: &'a Uri, method: Method) -> Request<'a> {
     let useragent = format!(
         "{}/{} ({}) {}",
         crate::APP_NAME,
@@ -39,27 +40,31 @@ fn new_request<'a>(
         env!("ACMED_TARGET"),
         env!("ACMED_HTTP_LIB_AGENT")
     );
-    let mut rb = Request::new(uri);
+    let mut rb = request::Request::new(uri);
     for file_name in root_certs.iter() {
         rb.root_cert_file_pem(&Path::new(file_name));
     }
-    rb.method(method);
+    rb.method(method.to_owned());
     rb.header("User-Agent", &useragent);
     // TODO: allow to configure the language
     rb.header("Accept-Language", "en-US,en;q=0.5");
-    rb
+    Request { r: rb, method, uri }
 }
 
-fn send_request(request: &Request) -> Result<(Response, String), Error> {
+fn send_request(cert: &Certificate, request: &Request) -> Result<(Response, String), Error> {
     let mut buffer = Vec::new();
-    let res = request.send(&mut buffer)?;
+    cert.https_throttle
+        .send(rate_limits::Request::HttpsRequest)
+        .unwrap();
+    cert.debug(&format!("{}: {}", request.method, request.uri));
+    let res = request.r.send(&mut buffer)?;
     let res_str = String::from_utf8(buffer)?;
     Ok((res, res_str))
 }
 
 fn send_request_retry(cert: &Certificate, request: &Request) -> Result<(Response, String), Error> {
     for _ in 0..crate::DEFAULT_HTTP_FAIL_NB_RETRY {
-        let (res, res_body) = send_request(request)?;
+        let (res, res_body) = send_request(cert, request)?;
         match check_response(&res, &res_body) {
             Ok(()) => {
                 return Ok((res, res_body));
@@ -110,14 +115,14 @@ fn post_jose_type(
     accept_type: &str,
 ) -> Result<(Response, String), Error> {
     let uri = url.parse::<Uri>()?;
-    let mut request = new_request(cert, root_certs, &uri, Method::POST);
-    request.header("Content-Type", CONTENT_TYPE_JOSE);
-    request.header("Content-Length", &data.len().to_string());
-    request.header("Accept", accept_type);
-    request.body(data);
+    let mut request = new_request(root_certs, &uri, Method::POST);
+    request.r.header("Content-Type", CONTENT_TYPE_JOSE);
+    request.r.header("Content-Length", &data.len().to_string());
+    request.r.header("Accept", accept_type);
+    request.r.body(data);
     let rstr = String::from_utf8_lossy(data);
     cert.trace(&format!("request body: {}", rstr));
-    let (res, res_body) = send_request(&request)?;
+    let (res, res_body) = send_request(cert, &request)?;
     cert.trace(&format!("response body: {}", res_body));
     Ok((res, res_body))
 }
@@ -287,8 +292,8 @@ pub fn get_directory(
     url: &str,
 ) -> Result<Directory, Error> {
     let uri = url.parse::<Uri>()?;
-    let mut request = new_request(cert, root_certs, &uri, Method::GET);
-    request.header("Accept", CONTENT_TYPE_JSON);
+    let mut request = new_request(root_certs, &uri, Method::GET);
+    request.r.header("Accept", CONTENT_TYPE_JSON);
     let (r, s) = send_request_retry(cert, &request)?;
     check_response(&r, &s)?;
     Directory::from_str(&s)
@@ -296,7 +301,7 @@ pub fn get_directory(
 
 pub fn get_nonce(cert: &Certificate, root_certs: &[String], url: &str) -> Result<String, Error> {
     let uri = url.parse::<Uri>()?;
-    let request = new_request(cert, root_certs, &uri, Method::HEAD);
+    let request = new_request(root_certs, &uri, Method::HEAD);
     let (res, res_body) = send_request_retry(cert, &request)?;
     check_response(&res, &res_body)?;
     nonce_from_response(cert, &res)